Updating to use AbstractEntity for model classes.
