+++ src/main/java/com/example/AdventureAppraisals/models/Destination.java
@@ -1,33 +1,9 @@
 package com.example.AdventureAppraisals.models;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 
 @Entity
-public class Destination {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    private String name;
+public class Destination extends AbstractEntity{
 
     public Destination() {}
-
-    public Destination(String name) {
-        this.name = name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
 }
